workers/requirements.txt
playwright==1.40.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
cryptography==41.0.7
